requirements.txt
# ============================================================
# AutoAnalyst Dependencies
# Compatible with Python 3.9+
# ============================================================

# --- Core Agent ---
langgraph>=0.2.60,<1.0
langchain>=0.3.14,<1.0
langchain-core>=0.3.29,<1.0
langchain-google-genai>=2.0.8,<3.0
langchain-openai>=0.3.0,<1.0

# --- Google Cloud ---
google-cloud-bigquery>=3.25.0,<4.0
google-cloud-bigquery-storage>=2.27.0,<3.0
google-auth>=2.37.0,<3.0
google-cloud-aiplatform>=1.74.0,<2.0
db-dtypes>=1.3.0,<2.0

# --- API Server ---
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.34.0,<1.0
pydantic>=2.10.0,<3.0
pydantic-settings>=2.7.0,<3.0
sse-starlette>=2.2.0,<3.0

# --- Dashboard ---
streamlit>=1.41.0,<2.0
plotly>=5.24.0,<6.0

# --- Data Processing ---
pandas>=2.0.0,<3.0
pyarrow>=15.0.0,<19.0
numpy>=1.24.0,<3.0

# --- Utilities ---
python-dotenv>=1.0.0,<2.0
structlog>=24.4.0,<25.0
pyyaml>=6.0,<7.0
tenacity>=9.0.0,<10.0
httpx>=0.27.0,<1.0

# --- Testing ---
pytest>=8.0.0,<9.0
pytest-asyncio>=0.25.0,<1.0
pytest-cov>=6.0.0,<7.0
